refactor: remove unused instrumentation file and update imports
- Deleted the unused instrumentation.ts file.
- Updated imports in various components to use the new sfcc structure.
- Added environment variable validation in sitemap and adjusted related components.
