+++ app/checkout/(main)/page.tsx
@@ -1,5 +1,5 @@
 import { getCart } from "@/lib/sfcc";
-import { getPathForCartCheckoutStep } from "@/lib/utils/checkout-helpers";
+import { getPathForCartCheckoutStep } from "@/lib/sfcc/utils";
 import { redirect } from "next/navigation";
 
 export default async function CheckoutPage() {
