+++ app/sitemap.ts
@@ -1,6 +1,6 @@
 import { getCollections, getProducts } from "lib/sfcc";
 import { getPages } from "lib/sfcc/content";
-import { baseUrl } from "lib/utils";
+import { baseUrl, validateEnvironmentVariables } from "lib/utils";
 import { MetadataRoute } from "next";
 
 type Route = {
@@ -11,6 +11,8 @@ type Route = {
 export const dynamic = "force-dynamic";
 
 export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
+  validateEnvironmentVariables();
+
   const routesMap = [""].map((route) => ({
     url: `${baseUrl}${route}`,
     lastModified: new Date().toISOString(),
