+++ components/checkout/payment-form.tsx
@@ -20,7 +20,7 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { billingAddressSchema, paymentFormSchema } from "@/lib/schemas";
-import { months, years } from "@/lib/utils/form-helpers";
+import { months, years } from "@/lib/sfcc/form-helpers";
 import clsx from "clsx";
 import { useState } from "react";
 import { useCart } from "../cart/cart-context";
