+++ lib/utils.ts
@@ -1,7 +1,6 @@
 import { clsx, type ClassValue } from "clsx";
 import { ReadonlyURLSearchParams } from "next/navigation";
 import { twMerge } from "tailwind-merge";
-import { z } from "zod";
 
 export const baseUrl = process.env.VERCEL_PROJECT_PRODUCTION_URL
   ? `https://${process.env.VERCEL_PROJECT_PRODUCTION_URL}`
@@ -24,32 +23,31 @@ export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
 }
 
-// Helper for returning the expected error state to actions instead of throwing.
-export const handleFormActionError = (error: unknown, defaultMessage: string) => {
-  return {
-    errors: {
-      formErrors: [(error as Error)?.message || defaultMessage],
-    },
-  };
-};
+export const validateEnvironmentVariables = () => {
+  const requiredEnvironmentVariables = [
+    "SITE_NAME",
+    "SFCC_CLIENT_ID",
+    "SFCC_ORGANIZATIONID",
+    "SFCC_SECRET",
+    "SFCC_SHORTCODE",
+    "SFCC_SITEID",
+    "SFCC_SANDBOX_DOMAIN",
+    "SFCC_OPENCOMMERCE_SHOP_API_ENDPOINT",
+    "SFCC_REVALIDATION_SECRET",
+  ];
+  const missingEnvironmentVariables = [] as string[];
 
-type PrefixedShape<T extends z.ZodObject<any>, P extends string> = {
-  [K in keyof T["shape"] as K extends string ? `${P}.${K}` : never]: T["shape"][K];
-};
+  requiredEnvironmentVariables.forEach((envVar) => {
+    if (!process.env[envVar]) {
+      missingEnvironmentVariables.push(envVar);
+    }
+  });
 
-// Creates a new Zod schema with all keys prefixed with the given string.
-export const prefixSchema = <T extends z.ZodObject<any>, P extends string>(
-  schema: T,
-  prefix: P
-): z.ZodObject<PrefixedShape<T, P>> => {
-  if (!prefix) return schema as z.ZodObject<PrefixedShape<T, P>>;
-
-  const shape = schema.shape;
-  const newShape = {} as PrefixedShape<T, P>;
-
-  for (const [key, value] of Object.entries(shape)) {
-    (newShape as any)[`${prefix}.${key}`] = value;
+  if (missingEnvironmentVariables.length) {
+    throw new Error(
+      `The following environment variables are missing. Your site will not work without them. Read more: https://vercel.com/docs/integrations/shopify#configure-environment-variables\n\n${missingEnvironmentVariables.join(
+        "\n"
+      )}\n`
+    );
   }
-
-  return z.object(newShape);
 };
